CD-70. SignOut started. Some bugs fixed

+++ CD/CD/Helper/FireBaseHelperMark.cs
@@ -15,7 +15,7 @@ namespace CD.Helper
         private readonly string Marks_Name = "Marks";
         readonly FirebaseClient firebase = new FirebaseClient(App.conf.firebase);
 
-        public async Task AddMark(Guid subjectID, string mark_name, decimal result, int weight, string category)
+        public async Task AddMark(Guid subjectID, string mark_name, double result, double weight, string category)
         {
             await firebase.Child(UserUID).Child(Marks_Name).PostAsync(new Mark()
             {
